Split continues — base UI components moved to core/ui_base/ .
Continues the separation started in the previous commit. The base UI blocks —
widgets, styling, tab scaffold, console input validation — and
domain/architectures are relocated into core/ui_base/ and core/domain/. Pure
move: classes keep their names and attributes; the main file now imports them.
Monolith: 26,551 → 25,101 lines. Per-module refactoring (decoupling, tests) is
deferred to separate work.
