+++ lora_trainer_gui.py
@@ -58,333 +58,11 @@ from fizgig_gui.core.config.constants import (
 )
 
 
-class _GUIWriter:
-    """Redirect stdout/stderr to the GUI log buffer when running under pythonw."""
-    def __init__(self, gui_ref, stream_name):
-        self._gui = gui_ref
-        self._name = stream_name
-    def write(self, text):
-        if text and text.strip():
-            try:
-                self._gui.master.after(0, self._gui._append_global_log,
-                                       f"[{self._name}] {text}")
-            except Exception:
-                pass
-    def flush(self):
-        pass
+from fizgig_gui.core.ui_base.widgets import _GUIWriter, ToolTip, CollapsibleFrame
 
-
-class ToolTip:
-    """Simple tooltip class for tkinter widgets"""
-    def __init__(self, widget, text):
-        self.widget = widget
-        self.text = text
-        self.tooltip_window = None
-        widget.bind("<Enter>", self.show_tooltip)
-        widget.bind("<Leave>", self.hide_tooltip)
-
-    def show_tooltip(self, event=None):
-        if self.tooltip_window:
-            return
-        bbox = self.widget.bbox("insert") if hasattr(self.widget, 'bbox') else None
-        bx, by = (bbox[0], bbox[1]) if bbox else (0, 0)
-        x = bx + self.widget.winfo_rootx() + 25
-        y = by + self.widget.winfo_rooty() + 25
-
-        self.tooltip_window = tw = tk.Toplevel(self.widget)
-        tw.wm_overrideredirect(True)
-        tw.wm_geometry(f"+{x}+{y}")
-
-        label = tk.Label(tw, text=self.text, justify=tk.LEFT,
-                        background=COLORS["bg_surface"], foreground=COLORS["text_primary"],
-                        relief=tk.SOLID, borderwidth=1,
-                        font=(FONT_FAMILY, 9), padx=8, pady=6)
-        label.pack()
-
-    def hide_tooltip(self, event=None):
-        if self.tooltip_window:
-            self.tooltip_window.destroy()
-            self.tooltip_window = None
-
-
-class CollapsibleFrame(tk.Frame):
-    """
-    A frame that can be collapsed/expanded with a header.
-
-    Features:
-    - Click header to toggle
-    - Arrow indicator (▶/▼)
-    - Optional badge showing field status
-    - Maintains child widget state when collapsed
-    - All child widgets remain accessible via parent.entries[]
-
-    Styled as a Start-tab-style card (bg_surface body, bordered outer frame).
-    """
-
-    def __init__(self, parent, title, default_expanded=True, badge_callback=None):
-        """
-        Args:
-            parent: Parent widget
-            title: Section title text
-            default_expanded: Whether section starts expanded
-            badge_callback: Optional function returning (filled, total) tuple for badge
-        """
-        super().__init__(
-            parent,
-            bg=COLORS["bg_surface"],
-            highlightbackground=COLORS["border"],
-            highlightthickness=1,
-            bd=0,
-        )
-        self.expanded = default_expanded
-        self.title = title
-        self.badge_callback = badge_callback
-
-        # Create header frame
-        self.header = tk.Frame(
-            self,
-            bg=COLORS["bg_header"],
-            cursor="hand2"
-        )
-        self.header.pack(fill=tk.X)
-
-        # Arrow indicator
-        self.arrow = tk.Label(
-            self.header,
-            text="▼" if self.expanded else "▶",
-            font=(FONT_FAMILY, 10),
-            fg=COLORS["text_secondary"],
-            bg=COLORS["bg_header"],
-            cursor="hand2"
-        )
-        self.arrow.pack(side=tk.LEFT, padx=(16, 10), pady=12)
-
-        # Title label — matches Start-tab card headers at 12pt bold
-        self.title_label = tk.Label(
-            self.header,
-            text=title,
-            font=(FONT_FAMILY, 12, "bold"),
-            fg=COLORS["text_primary"],
-            bg=COLORS["bg_header"],
-            cursor="hand2"
-        )
-        self.title_label.pack(side=tk.LEFT, pady=12)
-
-        # Badge label (shows filled/total fields)
-        self.badge = tk.Label(
-            self.header,
-            text="",
-            font=(FONT_FAMILY, 9),
-            fg=COLORS["text_secondary"],
-            bg=COLORS["bg_header"],
-            cursor="hand2"
-        )
-        self.badge.pack(side=tk.RIGHT, padx=(8, 16), pady=12)
-
-        # Content frame — bg_surface, padded from the card edge so children don't
-        # touch the border. Children grid into this directly.
-        self.content = tk.Frame(self, bg=COLORS["bg_surface"])
-        if self.expanded:
-            self.content.pack(fill=tk.BOTH, expand=True, padx=12, pady=(8, 12))
-
-        # Bind click events to all header elements
-        for widget in [self.header, self.arrow, self.title_label, self.badge]:
-            widget.bind("<Button-1>", self.toggle)
-            widget.bind("<Enter>", self._on_header_enter)
-            widget.bind("<Leave>", self._on_header_leave)
-
-    def _on_header_enter(self, event=None):
-        """Highlight header on hover"""
-        hover_color = COLORS["bg_hover"]
-        self.header.configure(bg=hover_color)
-        self.arrow.configure(bg=hover_color)
-        self.title_label.configure(bg=hover_color)
-        self.badge.configure(bg=hover_color)
-
-    def _on_header_leave(self, event=None):
-        """Reset header color on mouse leave"""
-        header_color = COLORS["bg_header"]
-        self.header.configure(bg=header_color)
-        self.arrow.configure(bg=header_color)
-        self.title_label.configure(bg=header_color)
-        self.badge.configure(bg=header_color)
-
-    def toggle(self, event=None):
-        """Toggle between expanded and collapsed states"""
-        if self.expanded:
-            self.content.pack_forget()
-            self.arrow.config(text="▶")
-        else:
-            self.content.pack(fill=tk.BOTH, expand=True, padx=12, pady=(8, 12))
-            self.arrow.config(text="▼")
-        self.expanded = not self.expanded
-
-    def expand(self):
-        """Expand the section if collapsed"""
-        if not self.expanded:
-            self.toggle()
-
-    def collapse(self):
-        """Collapse the section if expanded"""
-        if self.expanded:
-            self.toggle()
-
-    def update_badge(self, filled=None, total=None):
-        """Update the badge showing filled/total fields"""
-        if self.badge_callback:
-            filled, total = self.badge_callback()
-        if filled is not None and total is not None:
-            self.badge.config(text=f"[{filled}/{total}]")
-        else:
-            self.badge.config(text="")
-
-    def get_content_frame(self):
-        """Return the content frame where child widgets should be added"""
-        return self.content
-
-
-# Architecture configurations
-ARCHITECTURES = {
-    "Flux 2 Klein Base 9B": {
-        "train_script": "FizgigIndependent/src/fizgig/scripts/train.py",
-        "cache_latents_script": "FizgigIndependent/src/fizgig/scripts/cache_latents.py",
-        "cache_text_script": "FizgigIndependent/src/fizgig/scripts/cache_text.py",
-        "network_module": "fizgig.networks.lora_klein",
-        "use_fizgig_venv": True,
-        "timestep_sampling": "flux2_shift",
-        "discrete_flow_shift": None,
-        "weighting_scheme": "none",
-        "blocks_swap_max": 16,
-        "fp8_text_encoder_flag": "--fp8_text_encoder",
-        "uses_clip": False,
-        "uses_t5": False,
-        "uses_text_encoder": True,
-        "uses_model_type": False,
-        "uses_model_version": True,
-        "model_version": "klein-base-9b",
-        "vae_label": "AE Model (ae.safetensors from FLUX.2-dev — NOT the Diffusers subfolder VAE)",
-        "text_encoder_label": "Text Encoder (Qwen3-8B)",
-        "is_distilled": False,  # Recommended for training
-        "supports_weighting_scheme": False,  # Architecture uses hardcoded "none"
-        "supports_discrete_flow_shift": False,  # Uses flux2_shift automatic
-        # Sample generation settings
-        "supports_samples": True,
-        "sample_cfg_default": 4.5,
-        "sample_flow_shift_default": None,
-        "sample_steps_default": 40,  # Base is not step-distilled — BFL spec is ~50 steps; 40 balances quality vs sample time. Distilled path overrides to 4.
-        "sample_width_default": 768,
-        "sample_height_default": 768,
-        # Trailing tag on the default LoRA name, so a file says which family made it. Swapped
-        # when the family changes — see _apply_lora_name_suffix.
-        "lora_name_suffix": "k9b",
-    },
-    "Krea 2": {
-        # Krea 2 trains natively via fizgig.scripts.krea2_* (no accelerate launch — a
-        # single-process script). The command builders branch on "is_krea2" and ignore
-        # the Klein-shaped keys below; they're kept so start_training / the Samples tab /
-        # validate_inputs can read the same config shape without KeyErrors.
-        "train_script": "src/fizgig/scripts/krea2_train.py",
-        "cache_latents_script": "src/fizgig/scripts/krea2_cache_latents.py",
-        "cache_text_script": "src/fizgig/scripts/krea2_cache_text.py",
-        "is_krea2": True,
-        "network_module": "fizgig.networks.lora_klein",  # unused — krea2 trainer builds its own net
-        "use_fizgig_venv": True,
-        "timestep_sampling": "shift",
-        "discrete_flow_shift": 2.5,
-        "weighting_scheme": "none",
-        "blocks_swap_max": 26,  # Krea 2 SingleStreamDiT has 28 blocks; offloader caps at 28-2
-        "fp8_text_encoder_flag": None,
-        "uses_clip": False,
-        "uses_t5": False,
-        "uses_text_encoder": True,
-        "uses_model_type": False,
-        "uses_model_version": False,
-        "model_version": "krea-2",
-        "vae_label": "Qwen-Image VAE",
-        "text_encoder_label": "Qwen3-VL-4B",
-        "is_distilled": False,
-        "supports_weighting_scheme": False,
-        "supports_discrete_flow_shift": False,
-        # Sample generation: previews render on the fp8 Turbo (8-step; CFG optional —
-        # raise Sample CFG Scale above 1 with a negative prompt for guided previews).
-        "supports_samples": True,
-        "sample_cfg_default": 1.0,
-        "sample_flow_shift_default": None,
-        "sample_steps_default": 8,
-        "sample_width_default": 1024,
-        "sample_height_default": 1024,
-        "lora_name_suffix": "krea2",
-    },
-    "MiniMax H3": {
-        # MiniMax H3 trains natively via fizgig.scripts.minimax_* (single-process). The command
-        # builders branch on "is_minimax". Barebones IMAGE-ONLY: no samples, no preview, no
-        # per-image loss watch, no LoKR — the most minimal training surface. The Klein-shaped keys
-        # below are kept so start_training / Samples / validate_inputs read the same shape.
-        "train_script": "src/fizgig/scripts/minimax_train.py",
-        "cache_latents_script": "src/fizgig/scripts/minimax_cache_latents.py",
-        "cache_text_script": "src/fizgig/scripts/minimax_cache_text.py",
-        "is_minimax": True,
-        "network_module": "fizgig.networks.lora_klein",  # unused — minimax trainer builds its own net
-        "use_fizgig_venv": True,
-        "timestep_sampling": "shift",
-        "discrete_flow_shift": 12.0,   # H3 video sigma-shift (fixed; the Timestep section is hidden)
-        "weighting_scheme": "none",
-        "blocks_swap_max": 40,         # bnb NF4 blocks swap packed (uint8) — planner caps at 40 of 50
-        "fp8_text_encoder_flag": None,
-        "uses_clip": False,
-        "uses_t5": False,
-        "uses_text_encoder": True,
-        "uses_model_type": False,
-        "uses_model_version": False,
-        "model_version": "minimax-h3",
-        "vae_label": "MiniMax H3 Video VAE",
-        "text_encoder_label": "Qwen3-VL-32B",
-        "is_distilled": False,
-        "supports_weighting_scheme": False,
-        "supports_discrete_flow_shift": False,
-        # In-training previews render one still per prompt on the resident training DiT
-        # (latent_t=1, the training distribution). Per-epoch only, like Krea 2.
-        "supports_samples": True,
-        "sample_cfg_default": 1.0,
-        "sample_flow_shift_default": None,
-        # H3 samples run CFG-free on a fixed shift-12 schedule, exactly as every shipped ComfyUI
-        # workflow does (BasicGuider, no negative conditioning). Reusing the existing distilled
-        # flags greys out Negative Prompt and CFG Scale through the generic path rather than
-        # adding a MiniMax branch — nothing about them is editable on this family.
-        "sample_is_distilled": True,
-        "sample_cfg_fixed": True,
-        "sample_steps_default": 20,   # the reference pipeline default
-        # 768x768 (Peter, 17 Aug — down from the 11 Aug 1024): H3's native canvas is a 768
-        # short edge, and with clips-with-sound in the preview mix the smaller frame keeps
-        # those affordable too. 1024 is still one dropdown away.
-        "sample_width_default": 768,
-        "sample_height_default": 768,
-        "lora_name_suffix": "mmh3",
-    },
-}
-
-# Saved configs written before 3.6.1 carry the old label. Every lookup here is a .get() that
-# falls back to Klein, so without an alias a MiniMax preset would silently come back as a Klein
-# one - wrong family, no error. The alias points at the same config; _canon_arch maps it forward
-# so what the user then SEES is the current name.
-_ARCH_ALIASES = {"MiniMax H3 (experimental)": "MiniMax H3",
-                 "Krea 2 (experimental)": "Krea 2"}       # pre-rename saves (2026-07-28)
-for _old, _new in _ARCH_ALIASES.items():
-    ARCHITECTURES[_old] = ARCHITECTURES[_new]
-
-# Aliases are readable, not offerable: the dropdown lists current names only.
-ARCHITECTURE_LIST = [k for k in ARCHITECTURES if k not in _ARCH_ALIASES]
-
-
-def _canon_arch(name):
-    """Old label in, current label out."""
-    return _ARCH_ALIASES.get(name, name)
-
-# Every family suffix we recognise on a LoRA name. Used to swap ONE tag for another when the
-# model family changes — matching this set (never "the last underscore segment") is what stops
-# a name like portrait_v2 being mangled into portrait_krea2.
-LORA_NAME_SUFFIXES = {c["lora_name_suffix"] for c in ARCHITECTURES.values()
-                      if c.get("lora_name_suffix")}
+from fizgig_gui.core.domain.architectures import (
+    ARCHITECTURES, _ARCH_ALIASES, ARCHITECTURE_LIST, _canon_arch, LORA_NAME_SUFFIXES,
+)
 
 # Preview resolutions offered by BOTH the Samples tab and the live "Override next sample" panel.
 # They used to be two hardcoded lists that had drifted — the Samples tab reached 1536 while the
@@ -417,6 +95,9 @@ from fizgig_gui.core.config.presets import (
     ft_checkpoint_continuation, BUILT_IN_PRESETS, KREA2_BUILT_IN_PRESETS,
     SEED_TRAVEL_PRESETS, MINIMAX_BUILT_IN_PRESETS, _MM_DEFAULTS_KEY,
 )
+from fizgig_gui.core.ui_base.styling import StylingMixin
+from fizgig_gui.core.ui_base.tab_scaffold import TabScaffoldMixin
+from fizgig_gui.core.ui_base.console_validation import ConsoleValidationMixin
 
 from fizgig_gui.core.config.last_used import (
     DATASET_DIR, CACHE_DIR, OUTPUT_LORAS_DIR, LAST_USED_FILE,
@@ -439,7 +120,7 @@ from fizgig_gui.core.config.prefs import (
 
 from fizgig_gui.core.config.settings_map import SETTING_TO_PREF, PRESETS
 
-class LoRATrainerGUI:
+class LoRATrainerGUI(StylingMixin, TabScaffoldMixin, ConsoleValidationMixin):
     def __init__(self, master):
         self.master = master
         master.title("Fizgig — Klein 9B & Krea 2 LoRA Studio")
@@ -1738,654 +1419,6 @@ class LoRATrainerGUI:
             initialdir=self._pref_initialdir("input_ref_dir"))
         if path:
             self.sample_ref_image_var.set(path)
-
-    def setup_styles(self):
-        """Set up styles for refined dark theme (Fizgig Visual Style Guide)"""
-        style = ttk.Style()
-        style.theme_use("clam")
-
-        # Base styles with new palette
-        style.configure(".",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-        style.configure("TFrame", background=COLORS["bg_deep"])
-        style.configure("TLabel",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-
-        # Surface style for cards/panels
-        style.configure("Surface.TFrame", background=COLORS["bg_surface"])
-
-        # Green render-progress bar (Repair Studio) — matches the Start button's green.
-        style.configure("Green.Horizontal.TProgressbar",
-                        background="#2E8B57", troughcolor=COLORS["bg_deep"],
-                        bordercolor=COLORS["bg_deep"],
-                        lightcolor="#2E8B57", darkcolor="#2E8B57")
-
-        # Collapsible header style
-        style.configure("CollapsibleHeader.TFrame", background=COLORS["bg_header"])
-
-        # Section header label style
-        style.configure("SectionHeader.TLabel",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 12, "bold")
-        )
-
-        # Secondary label style (for field labels)
-        style.configure("Secondary.TLabel",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_secondary"],
-            font=(FONT_FAMILY, 10)
-        )
-
-        # Default button (Secondary style)
-        style.configure(
-            "TButton",
-            background=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            bordercolor=COLORS["border"],
-            borderwidth=1,
-            focusthickness=2,
-            focuscolor=COLORS["accent"],
-            padding=[16, 8],
-            font=(FONT_FAMILY, 10, "bold")
-        )
-        style.map(
-            "TButton",
-            background=[("active", COLORS["bg_hover"]), ("pressed", COLORS["bg_hover"])],
-            foreground=[("active", COLORS["text_primary"]), ("pressed", COLORS["text_primary"])]
-        )
-
-        # Primary button (accent color)
-        style.configure(
-            "Primary.TButton",
-            background=COLORS["accent"],
-            foreground="white",
-            bordercolor=COLORS["accent"],
-            borderwidth=1,
-            focusthickness=2,
-            focuscolor=COLORS["accent_hover"],
-            padding=[16, 8],
-            font=(FONT_FAMILY, 10, "bold")
-        )
-        style.map(
-            "Primary.TButton",
-            background=[("active", COLORS["accent_hover"]), ("pressed", COLORS["accent_hover"])],
-            foreground=[("active", "white"), ("pressed", "white")]
-        )
-
-        # Danger button (for stop actions)
-        style.configure(
-            "Danger.TButton",
-            background=COLORS["error"],
-            foreground="white",
-            bordercolor=COLORS["error"],
-            borderwidth=1,
-            focusthickness=2,
-            focuscolor=COLORS["error"],
-            padding=[16, 8],
-            font=(FONT_FAMILY, 10, "bold")
-        )
-        style.map(
-            "Danger.TButton",
-            background=[("active", "#DC2626"), ("pressed", "#DC2626")],
-            foreground=[("active", "white"), ("pressed", "white")]
-        )
-
-        # Checkbutton
-        style.configure("TCheckbutton",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("TCheckbutton",
-            background=[("active", COLORS["bg_deep"])],
-            foreground=[("active", COLORS["text_primary"])]
-        )
-        style.configure("TRadiobutton",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("TRadiobutton",
-            background=[("active", COLORS["bg_deep"])],
-            foreground=[("active", COLORS["text_primary"])]
-        )
-        style.configure("Surface.TRadiobutton",
-            background=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("Surface.TRadiobutton",
-            background=[("active", COLORS["bg_surface"])],
-            foreground=[("active", COLORS["text_primary"])]
-        )
-
-        # Surface checkbutton (for use inside collapsible sections)
-        style.configure("Surface.TCheckbutton",
-            background=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("Surface.TCheckbutton",
-            background=[("active", COLORS["bg_surface"])],
-            foreground=[("active", COLORS["text_primary"])]
-        )
-
-        # Notebook (tabs)
-        style.configure("TNotebook",
-            background=COLORS["bg_deep"],
-            borderwidth=0
-        )
-        style.configure("TNotebook.Tab",
-            background=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            padding=[12, 6],
-            font=(FONT_FAMILY, 11, "bold")
-        )
-        style.map("TNotebook.Tab",
-            background=[("selected", COLORS["accent"])],
-            foreground=[("selected", "white")]
-        )
-
-        # Entry field — explicit insert cursor settings so the caret is visible on click
-        style.configure(
-            "TEntry",
-            fieldbackground=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            bordercolor=COLORS["border"],
-            insertcolor=COLORS["accent"],
-            insertwidth=2,
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("TEntry",
-            fieldbackground=[("focus", ACTIVE_ENTRY_BG)],
-            foreground=[("focus", ACTIVE_ENTRY_FG)],
-            bordercolor=[("focus", COLORS["border_focus"])],
-            insertcolor=[("focus", COLORS["accent"])],
-        )
-
-        # Combobox — explicit insert cursor settings so editable comboboxes show a caret
-        style.configure(
-            "TCombobox",
-            fieldbackground=COLORS["bg_surface"],
-            background=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            bordercolor=COLORS["border"],
-            arrowcolor=COLORS["text_secondary"],
-            insertcolor=COLORS["accent"],
-            insertwidth=2,
-            font=(FONT_FAMILY, 10)
-        )
-        style.map("TCombobox",
-            fieldbackground=[("focus", ACTIVE_ENTRY_BG), ("readonly", COLORS["bg_surface"]), ("!disabled", COLORS["bg_surface"])],
-            foreground=[("focus", ACTIVE_ENTRY_FG), ("readonly", COLORS["text_primary"]), ("!disabled", COLORS["text_primary"])],
-            selectbackground=[("readonly", COLORS["bg_surface"]), ("!disabled", COLORS["bg_surface"])],
-            selectforeground=[("readonly", COLORS["text_primary"]), ("!disabled", COLORS["text_primary"])],
-            bordercolor=[("focus", COLORS["border_focus"])]
-        )
-
-        # Treeview (the Metadata tab's custom-fields table). Nothing styled this before it —
-        # "clam" falls back to a stock white row background with no matching foreground, so
-        # rows rendered as near-invisible light-grey-on-white against an otherwise dark app.
-        style.configure(
-            "Treeview",
-            background=COLORS["bg_surface"],
-            fieldbackground=COLORS["bg_surface"],
-            foreground=COLORS["text_primary"],
-            bordercolor=COLORS["border"],
-            font=(FONT_FAMILY, 10),
-            rowheight=24,
-        )
-        style.map("Treeview",
-            background=[("selected", COLORS["accent"])],
-            foreground=[("selected", "white")]
-        )
-        style.configure(
-            "Treeview.Heading",
-            background=COLORS["bg_header"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10, "bold"),
-        )
-        style.map("Treeview.Heading",
-            background=[("active", COLORS["bg_hover"])]
-        )
-
-        # Scrollbar. `background` is the thumb — the part you drag — and it used to be bg_header
-        # on a bg_deep trough, which is 1.06:1. Not "subtle": indistinguishable from the track,
-        # so on a tall tab there was no visible clue that the panel scrolled at all.
-        #
-        # The accent blue, same as the selected tab. Colour lives in COLORS as its own entry so
-        # the scrollbar can be retuned without touching what "accent" means everywhere else.
-        for _orient in ("Vertical", "Horizontal"):
-            style.configure(
-                f"{_orient}.TScrollbar",
-                background=COLORS["scrollbar_thumb"],
-                troughcolor=COLORS["bg_deep"],
-                bordercolor=COLORS["border"],
-                arrowcolor=COLORS["text_primary"],
-                darkcolor=COLORS["bg_deep"],
-                lightcolor=COLORS["bg_deep"],
-                width=12
-            )
-            style.map(
-                f"{_orient}.TScrollbar",
-                background=[("active", COLORS["scrollbar_thumb_hover"]),
-                            ("pressed", COLORS["scrollbar_thumb_hover"])]
-            )
-
-        # LabelFrame
-        style.configure("TLabelframe",
-            background=COLORS["bg_deep"],
-            bordercolor=COLORS["border"]
-        )
-        style.configure("TLabelframe.Label",
-            background=COLORS["bg_deep"],
-            foreground=COLORS["text_primary"],
-            font=(FONT_FAMILY, 10, "bold")
-        )
-
-    def create_scrollable_frame(self, parent):
-        """Create a scrollable frame within a parent widget"""
-        # Create a canvas
-        canvas = tk.Canvas(parent, bg=BG_COLOR, highlightthickness=0)
-        scrollbar = ttk.Scrollbar(parent, orient="vertical", command=canvas.yview)
-        scrollable_frame = ttk.Frame(canvas)
-
-        # Configure the scrollable frame
-        scrollable_frame.bind(
-            "<Configure>",
-            lambda e: canvas.configure(scrollregion=canvas.bbox("all"))
-        )
-
-        # Create window inside canvas
-        canvas_window = canvas.create_window((0, 0), window=scrollable_frame, anchor="nw")
-
-        # Configure canvas to expand horizontally
-        def configure_canvas(event):
-            canvas.itemconfig(canvas_window, width=event.width)
-
-        canvas.bind("<Configure>", configure_canvas)
-        canvas.configure(yscrollcommand=scrollbar.set)
-
-        # Pack the canvas and scrollbar
-        canvas.pack(side="left", fill="both", expand=True)
-        scrollbar.pack(side="right", fill="y")
-
-        # Mousewheel: handled by the global router (_route_mousewheel) — it finds this
-        # canvas through the pointer's ancestry, so no per-tab bind_all is needed (the old
-        # Enter/Leave bind_all dance is what made scrollable areas fight each other).
-
-        return scrollable_frame, canvas
-
-    def _add_tab_banner(self, parent, title, subtitle):
-        """Start-tab-style banner (22pt title + 11pt subtitle on bg_deep).
-        Packs into `parent`; returns the banner frame in case the caller wants to tweak it."""
-        banner = tk.Frame(parent, bg=COLORS["bg_deep"])
-        banner.pack(fill=tk.X, padx=36, pady=(28, 20))
-        tk.Label(banner, text=title,
-                 font=(FONT_FAMILY, 22, "bold"),
-                 fg=COLORS["text_primary"], bg=COLORS["bg_deep"]).pack(anchor=tk.W)
-        if subtitle:
-            # Colour only — 11pt is already comfortable, and this passed contrast before. Moved
-            # onto text_explain so every piece of explanatory copy is one colour, rather than the
-            # subtitle being a shade apart from the card descriptions directly beneath it.
-            tk.Label(banner, text=subtitle,
-                     font=(FONT_FAMILY, 11),
-                     fg=COLORS["text_explain"], bg=COLORS["bg_deep"],
-                     wraplength=1050, justify=tk.LEFT).pack(anchor=tk.W, pady=(4, 0))
-        return banner
-
-    def _add_youtube_help_button(self, parent, tab_key="start", prominent=False):
-        """Add a 'Tutorial' button at the bottom of a tab's outer frame.
-
-        Labelled just "Tutorial" rather than "Get help on YouTube": the play glyph and the red
-        already say where it goes, and "get help" reads as troubleshooting when these are walk-
-        throughs. Shorter also keeps the Start tab's button row on one line.
-
-        `tab_key` selects the URL from help.json's `youtube_urls` dict.
-        `prominent=True` uses a larger button and the Start tab's extra buttons alongside.
-        """
-        # Used only when help.json is missing or unreadable — point it at the real guide
-        # rather than the old joke URL, since that path fires on a genuine error.
-        fallback = "https://www.youtube.com/watch?v=yrz0l6URGGk"
-        try:
-            with open(HELP_FILE, "r", encoding="utf-8") as f:
-                urls = json.load(f).get("youtube_urls", {})
-            url = urls.get(tab_key, fallback)
-        except Exception:
-            url = fallback
-        btn_frame = tk.Frame(parent, bg=COLORS["bg_deep"])
-        btn_frame.pack(fill=tk.X, padx=36, pady=(16, 16) if prominent else (8, 8))
-        row = tk.Frame(btn_frame, bg=COLORS["bg_deep"])
-        row.pack(anchor=tk.W)
-        btn = tk.Button(
-            row, text="\u25b6  Tutorial",
-            font=(FONT_FAMILY, 12 if prominent else 10, "bold"),
-            fg="#FFFFFF", bg="#CC0000", activeforeground="#FFFFFF", activebackground="#990000",
-            relief="flat", bd=0, padx=20 if prominent else 16,
-            pady=6, cursor="hand2",
-            command=lambda: __import__("webbrowser").open(url),
-        )
-        btn.pack(side=tk.LEFT)
-        if prominent:
-            coffee = tk.Button(
-                row, text="\u2615  Buy me a coffee",
-                font=(FONT_FAMILY, 12, "bold"),
-                fg="#000000", bg="#FFDD00", activeforeground="#000000", activebackground="#E5C700",
-                relief="flat", bd=0, padx=20, pady=6, cursor="hand2",
-                command=lambda: __import__("webbrowser").open(
-                    "https://buymeacoffee.com/lorasandlenses"),
-            )
-            coffee.pack(side=tk.LEFT, padx=(12, 0))
-            # Shown on a pod too, deliberately: the Preferences card already tells a pod user they
-            # are on one, and a user who is happy renting is exactly who might send the link on.
-            # RunPod's own violet, so it reads as theirs rather than as another Fizgig action.
-            runpod = tk.Button(
-                row, text="⚡  Deploy on RunPod",
-                font=(FONT_FAMILY, 12, "bold"),
-                fg="#FFFFFF", bg="#7C3AED", activeforeground="#FFFFFF", activebackground="#6D28D9",
-                relief="flat", bd=0, padx=20, pady=6, cursor="hand2",
-                command=lambda: __import__("webbrowser").open(self._runpod_deploy_url()),
-            )
-            runpod.pack(side=tk.LEFT, padx=(12, 0))
-            # width sized for "⬆ Update Available" so the label can flip to it after the
-            # startup update check without shifting the whole button row.
-            about = tk.Button(
-                row, text="About", width=16,
-                font=(FONT_FAMILY, 12, "bold"),
-                fg="#FFFFFF", bg=COLORS["accent"],
-                activeforeground="#FFFFFF", activebackground=COLORS["accent_hover"],
-                relief="flat", bd=0, padx=20, pady=6, cursor="hand2",
-                command=self._open_about_dialog,
-            )
-            about.pack(side=tk.LEFT, padx=(12, 0))
-            self._about_btn = about
-
-    # ------------------------------------------------------------
-    # Update check (startup + manual from the About dialog)
-    # ------------------------------------------------------------
-
-    def _start_update_check(self, on_done=None):
-        """Run the git-based release check on a daemon thread; marshal the result back to
-        the Tk thread. Silent by design: 'unknown' (offline, no origin) shows nothing —
-        never a false nag, never an error popup."""
-        def _worker():
-            status, info = _check_for_update()
-            self.master.after(0, lambda: self._apply_update_status(status, info, on_done))
-        threading.Thread(target=_worker, daemon=True).start()
-
-    def _apply_update_status(self, status, info, on_done=None):
-        if status == "update_available":
-            self._update_info = info                     # the newer tag, e.g. "v3.1.2"
-            btn = getattr(self, "_about_btn", None)
-            if btn is not None:
-                try:
-                    btn.config(text="⬆ Update Available",
-                               bg=COLORS["warning"], activebackground="#D97706",
-                               fg="#1E2530", activeforeground="#1E2530")
-                except Exception:
-                    pass
-        elif status == "up_to_date":
-            self._update_info = None
-            btn = getattr(self, "_about_btn", None)
-            if btn is not None:
-                try:
-                    btn.config(text="About", bg=COLORS["accent"],
-                               activebackground=COLORS["accent_hover"],
-                               fg="#FFFFFF", activeforeground="#FFFFFF")
-                except Exception:
-                    pass
-        if on_done:
-            try:
-                on_done(status, info)
-            except Exception:
-                pass
-
-    def _open_about_dialog(self):
-        """A small, personal About window: who made Fizgig, why, and a no-pressure
-        nudge to the tip jar."""
-        import webbrowser
-        win = tk.Toplevel(self.master)
-        win.title("About Fizgig")
-        win.configure(bg=COLORS["bg_deep"])
-        win.transient(self.master)
-        win.resizable(False, False)
-        try:
-            win.grab_set()
-        except Exception:
-            pass
-
-        pad = tk.Frame(win, bg=COLORS["bg_deep"])
-        pad.pack(fill=tk.BOTH, expand=True, padx=28, pady=24)
-        WRAP = 460
-
-        def heading(text, size=20, fg=None, pady=(0, 2)):
-            tk.Label(pad, text=text, font=(FONT_FAMILY, size, "bold"),
-                     fg=fg or COLORS["text_primary"], bg=COLORS["bg_deep"]).pack(anchor=tk.W, pady=pady)
-
-        def para(text, fg=None, italic=False, pady=(0, 10)):
-            tk.Label(pad, text=text, font=(FONT_FAMILY, 10, "italic" if italic else "normal"),
-                     fg=fg or COLORS["text_secondary"], bg=COLORS["bg_deep"],
-                     wraplength=WRAP, justify=tk.LEFT).pack(anchor=tk.W, pady=pady)
-
-        def link(text, url, pady=(0, 2)):
-            lbl = tk.Label(pad, text=text, font=(FONT_FAMILY, 10, "underline"),
-                           fg=COLORS["accent_hover"], bg=COLORS["bg_deep"], cursor="hand2")
-            lbl.pack(anchor=tk.W, pady=pady)
-            lbl.bind("<Button-1>", lambda e, u=url: webbrowser.open(u))
-
-        heading("Fizgig", 22)
-        tk.Label(pad, text="Klein 9B & Krea 2 LoRA Studio — by Peter Neill",
-                 font=(FONT_FAMILY, 11), fg=COLORS["text_explain"],
-                 bg=COLORS["bg_deep"]).pack(anchor=tk.W, pady=(0, 4))
-        tk.Label(pad, text=f"Version {_git_describe_version() or 'unknown'}",
-                 font=(FONT_FAMILY, 9), fg=COLORS["text_muted"],
-                 bg=COLORS["bg_deep"]).pack(anchor=tk.W, pady=(0, 2))
-        # Always shown — how THIS install updates, whether or not one is pending right now.
-        _how = ("On RunPod, Fizgig updates itself every time you restart the pod — just stop "
-                "and start it to get the latest."
-                if _running_on_pod() else
-                ("To update: close Fizgig and run update_fizgig_rocm.bat."
-                 if os.environ.get("FIZGIG_GPU_BACKEND", "").lower() == "rocm" else
-                 "To update: close Fizgig and run update_fizgig.bat."))
-        tk.Label(pad, text=_how, font=(FONT_FAMILY, 9), fg=COLORS["text_muted"],
-                 bg=COLORS["bg_deep"], wraplength=WRAP, justify=tk.LEFT).pack(
-            anchor=tk.W, pady=(0, 10))
-
-        # Update banner — its own frame so the "Check for updates" button below can rebuild it
-        # live without reopening the dialog. Rendered per the current self._update_info.
-        update_box = tk.Frame(pad, bg=COLORS["bg_deep"])
-        update_box.pack(fill=tk.X, pady=(0, 8))
-
-        def _render_update_box():
-            for w in update_box.winfo_children():
-                w.destroy()
-            tag = getattr(self, "_update_info", None)
-            if not tag:
-                return
-            card = tk.Frame(update_box, bg=COLORS["accent_subtle"],
-                            highlightbackground=COLORS["warning"], highlightthickness=1)
-            card.pack(fill=tk.X)
-            tk.Label(card, text=f"⬆  Fizgig {tag} is available",
-                     font=(FONT_FAMILY, 11, "bold"), fg=COLORS["text_primary"],
-                     bg=COLORS["accent_subtle"]).pack(anchor=tk.W, padx=12, pady=(10, 0))
-            _upd = ("Restart the pod (stop and start it) to update."
-                    if _running_on_pod() else
-                    ("Close Fizgig and run update_fizgig_rocm.bat to update."
-                     if os.environ.get("FIZGIG_GPU_BACKEND", "").lower() == "rocm" else
-                     "Close Fizgig and run update_fizgig.bat to update."))
-            tk.Label(card, text=f"You're on {_git_describe_version() or 'an older build'}. {_upd}",
-                     font=(FONT_FAMILY, 9), fg=COLORS["text_explain"], bg=COLORS["accent_subtle"],
-                     wraplength=WRAP - 24, justify=tk.LEFT).pack(anchor=tk.W, padx=12, pady=(2, 8))
-            notes = tk.Label(card, text="📖  Read the release notes",
-                             font=(FONT_FAMILY, 10, "underline"), fg=COLORS["accent_hover"],
-                             bg=COLORS["accent_subtle"], cursor="hand2")
-            notes.pack(anchor=tk.W, padx=12, pady=(0, 10))
-            notes.bind("<Button-1>", lambda e: webbrowser.open(
-                "https://github.com/shootthesound/Fizgig/releases/latest"))
-
-        _render_update_box()
-
-        para("By trade I'm a photographer and videographer — mostly live music, portraits, and a "
-             "bit of teaching — and an AI tinkerer by night. I build a lot of open-source tooling "
-             "for ComfyUI and Klein/Flux workflows.")
-        link("Photography — shootthesound.com", "https://shootthesound.com")
-        link("Code & ComfyUI nodes — github.com/shootthesound", "https://github.com/shootthesound", pady=(0, 4))
-        para("(Realtime-LoRA, LongLook, Angelo, mesh and a couple of dozen more — Fizgig grew out of that world.)",
-             fg=COLORS["text_muted"], pady=(0, 16))
-
-        tk.Frame(pad, bg=COLORS["border"], height=1).pack(fill=tk.X, pady=(0, 16))
-
-        para("A quiet note: a lot of this got built in the small hours. The last year has been a hard "
-             "one for our family — one of my children has been facing some serious health challenges — "
-             "and honestly, losing myself in making and obsessing over tools like this is how I carve out "
-             "a little headspace. It keeps my hands busy and my head somewhere steady.", italic=True)
-
-        para("Fizgig is free and always will be. If it's useful to you and you'd like to drop a coffee in "
-             "the tip jar, it genuinely means a lot right now — but it's in no way an obligation. Using it "
-             "and enjoying it is more than enough. Thank you for being here.",
-             fg=COLORS["text_secondary"], pady=(0, 18))
-
-        btn_row = tk.Frame(pad, bg=COLORS["bg_deep"])
-        btn_row.pack(anchor=tk.W)
-        # No coffee button here on purpose — it's already on the Start tab, so
-        # repeating it in the popup would feel pushy. The note above is enough.
-        tk.Button(btn_row, text="Close", font=(FONT_FAMILY, 11),
-                  fg=COLORS["text_primary"], bg=COLORS["bg_surface"],
-                  activeforeground=COLORS["text_primary"], activebackground=COLORS["border"],
-                  relief="flat", bd=0, padx=18, pady=8, cursor="hand2",
-                  command=win.destroy).pack(side=tk.LEFT)
-
-        check_btn = tk.Button(btn_row, text="Check for updates", font=(FONT_FAMILY, 11),
-                              fg=COLORS["text_primary"], bg=COLORS["bg_surface"],
-                              activeforeground=COLORS["text_primary"],
-                              activebackground=COLORS["border"],
-                              relief="flat", bd=0, padx=18, pady=8, cursor="hand2")
-        check_btn.pack(side=tk.LEFT, padx=(12, 0))
-
-        def _manual_check():
-            if not win.winfo_exists():
-                return
-            check_btn.config(text="Checking…", state=tk.DISABLED)
-
-            def _done(status, info):
-                if not win.winfo_exists():
-                    return
-                _render_update_box()
-                check_btn.config(
-                    state=tk.NORMAL,
-                    text=("Up to date ✓" if status == "up_to_date"
-                          else "Couldn't check" if status == "unknown"
-                          else "Check for updates"))
-            self._start_update_check(on_done=_done)
-
-        check_btn.config(command=_manual_check)
-
-        win.update_idletasks()
-        # Centre over the main window.
-        try:
-            px = self.master.winfo_rootx() + (self.master.winfo_width() - win.winfo_width()) // 2
-            py = self.master.winfo_rooty() + (self.master.winfo_height() - win.winfo_height()) // 3
-            win.geometry(f"+{max(0, px)}+{max(0, py)}")
-        except Exception:
-            pass
-
-    def _start_section_card(self, parent, title, description=None, accent_border=False):
-        """Start-tab-style surface card with an optional description line.
-        Returns the inner content frame (caller packs/grids its widgets into it).
-        The card is packed into `parent` with horizontal padding matching the banner."""
-        outer = tk.Frame(parent, bg=COLORS["bg_deep"])
-        outer.pack(fill=tk.X, padx=36, pady=(0, 16))
-
-        card = tk.Frame(outer, bg=COLORS["bg_surface"],
-                        highlightbackground=COLORS["accent"] if accent_border else COLORS["border"],
-                        highlightthickness=1, bd=0)
-        card.pack(fill=tk.X)
-
-        if title:
-            tk.Label(card, text=title,
-                     font=(FONT_FAMILY, 12, "bold"),
-                     fg=COLORS["text_primary"], bg=COLORS["bg_surface"]).pack(
-                anchor=tk.W, padx=20, pady=(16, 2 if description else 10)
-            )
-        desc_label = None
-        if description:
-            desc_label = tk.Label(card, text=description,
-                                  font=(FONT_FAMILY, 10),
-                                  fg=COLORS["text_explain"], bg=COLORS["bg_surface"],
-                                  wraplength=760, justify=tk.LEFT)
-            desc_label.pack(anchor=tk.W, padx=20, pady=(0, 10))
-
-        content = tk.Frame(card, bg=COLORS["bg_surface"])
-        content.pack(fill=tk.X, padx=20, pady=(0, 16))
-        # Stashed so a caller can reword the description later (the Samples tab retitles its
-        # cards per model family). None when the card was built without one.
-        content._desc_label = desc_label
-        return content
-
-    def _add_field_to_section(self, parent, key, label_text, input_type, row):
-        """Helper method to add a field to a section (collapsible or regular frame)"""
-        # Create label
-        label = tk.Label(
-            parent,
-            text=f"{label_text}:",
-            font=(FONT_FAMILY, 10),
-            fg=COLORS["text_secondary"],
-            bg=COLORS["bg_surface"]
-        )
-        label.grid(row=row, column=0, sticky=tk.W, padx=(12, 8), pady=4)
-        self.labels[key] = label
-
-        # Create entry/combobox based on type
-        if input_type == "dropdown":
-            if key == "MODEL_TYPE":
-                arch = self.settings["ARCHITECTURE"]
-                arch_config = ARCHITECTURES.get(arch, ARCHITECTURES["Flux 2 Klein Base 9B"])
-                model_types = arch_config.get("model_types", ["t2v-14B", "i2v-14B"])
-                var = tk.StringVar(value=self.settings[key])
-                self.entries[key] = ttk.Combobox(parent, textvariable=var, values=model_types, state="readonly", width=38)
-                if self.settings[key] in model_types:
-                    self.entries[key].current(model_types.index(self.settings[key]))
-                else:
-                    self.entries[key].current(0)
-            elif key == "OPTIMIZER_TYPE":
-                var = tk.StringVar(value=self.settings[key])
-                self.entries[key] = ttk.Combobox(parent, textvariable=var, values=self.optimizer_types, state="readonly", width=38)
-                # Saved settings may carry a name that's no longer offered (e.g. the removed
-                # adafactor/prodigy/came) — fall back to the default instead of crashing.
-                if self.settings[key] in self.optimizer_types:
-                    self.entries[key].current(self.optimizer_types.index(self.settings[key]))
-                else:
-                    self.entries[key].set("adamw8bit")
-            elif key == "LR_SCHEDULER":
-                lr_scheduler_options = ["constant", "constant_with_warmup", "cosine", "cosine_with_restarts", "linear", "polynomial"]
-                self.lr_scheduler_var = tk.StringVar(value=self.settings["LR_SCHEDULER"])
-                self.entries[key] = ttk.Combobox(parent, textvariable=self.lr_scheduler_var, values=lr_scheduler_options, state="readonly", width=38)
-        else:
-            # Check if this entry should be bound to a shared pref var
-            pref_key = SETTING_TO_PREF.get(key)
-            if pref_key and pref_key in self.prefs_vars:
-                self.entries[key] = ttk.Entry(parent, width=40, textvariable=self.prefs_vars[pref_key])
-            else:
-                self.entries[key] = ttk.Entry(parent, width=40)
-                self.entries[key].insert(0, str(self.settings.get(key, "")))
-
-        self.entries[key].grid(row=row, column=1, sticky=tk.EW, padx=5, pady=4)
-
-        # Create browse button if needed
-        browse_btn = None
-        if input_type in ["file", "directory"]:
-            browse_btn = ttk.Button(parent, text="Browse", command=lambda k=key, t=input_type: self.browse_file(k, t))
-            browse_btn.grid(row=row, column=2, sticky=tk.W, padx=(5, 12), pady=4)
-
-        # Store row info for show/hide functionality
-        self.rows[key] = {"row": row, "label": label, "entry": self.entries[key], "browse": browse_btn, "parent": parent}
 
     def create_start_tab(self):
         """Welcome screen + Training image folder picker — the single source of truth shared with
@@ -23784,489 +22817,6 @@ class LoRATrainerGUI:
                 toml_lines.append("is_reg = true")
 
             return dataset_name, "\n".join(toml_lines) + "\n"
-
-    def show_context_menu(self, event):
-        """Show context menu on right-click"""
-        try:
-            self.context_menu.tk_popup(event.x_root, event.y_root)
-        finally:
-            self.context_menu.grab_release()
-
-    def copy_selected_text(self):
-        """Copy selected text to clipboard"""
-        if self.console_output.selection_get():
-            self.master.clipboard_clear()
-            self.master.clipboard_append(self.console_output.selection_get())
-
-    def browse_directory(self, setting_name):
-        path = filedialog.askdirectory()
-        if path:
-            self.entries[setting_name].delete(0, tk.END)
-            self.entries[setting_name].insert(0, path)
-
-    def on_mousewheel(self, event):
-        """Handle scroll event"""
-        if self.console_output.yview()[1] < 1.0:
-            self.user_scrolled = True
-        else:
-            self.user_scrolled = False
-
-    def update_console(self, line):
-        """Update training console — only auto-scroll if user was already at the bottom.
-        Uses the widget's own yview() position as the authoritative signal; the older
-        self.user_scrolled flag sometimes got out of sync with actual widget state."""
-        self._append_global_log(line)
-        try:
-            at_bottom = self.console_output.yview()[1] >= 0.999
-        except Exception:
-            at_bottom = True
-        self.console_output.configure(state="normal")
-        self.console_output.insert(tk.END, line)
-        if at_bottom:
-            self.console_output.see(tk.END)
-        self.console_output.configure(state="disabled")
-
-        # Detect CUDA OOM. A Krea 2 training-step OOM is best fixed by the 4-bit (NF4) base (it
-        # frees far more than swap); otherwise suggest more block swap. (Preview OOMs are caught in
-        # the trainer, auto-disable previews, and don't print this literal — so this only fires on a
-        # genuine training-step OOM.)
-        # The preview resolution ladder settled somewhere (trainer prints this per
-        # downgrade): write it back into the Samples tab so the NEXT run starts at the size
-        # that fit, instead of re-walking the ladder from the configured resolution. It
-        # persists exactly like a hand edit of Width/Height.
-        if "[preview] resolution settled:" in line:
-            import re as _re_res
-            _m = _re_res.search(r"resolution settled: (\d+)x(\d+)", line)
-            if _m and hasattr(self, "sample_width_var"):
-                self.sample_width_var.set(_m.group(1))
-                self.sample_height_var.set(_m.group(2))
-                self.update_console(f"[samples] preview resolution saved as the new "
-                                    f"default: {_m.group(1)}x{_m.group(2)} — future runs "
-                                    f"start there.\n")
-
-        if "CUDA out of memory" in line or "OutOfMemoryError" in line:
-            if not getattr(self, "_oom_warning_shown", False):
-                self._oom_warning_shown = True
-                # Read state WITHOUT side effects: _parse_blocks_swap() on "Auto" runs the
-                # auto strategy (GPU probe + can flip the 4-bit toggle) — it mutated the
-                # training config as a side effect of reporting an error, and did so BEFORE
-                # reading the nf4 flag it then tested, hiding the Krea 2 advice. Mid-run the
-                # probe is meaningless anyway (the trainer holds the VRAM).
-                _raw_swap = self.entries["BLOCKS_SWAP"].get().strip()
-                if _raw_swap.lower().startswith("auto"):
-                    current_swap, _swap_disp = 0, "Auto"
-                else:
-                    import re as _re
-                    _m = _re.match(r"\d+", _raw_swap)
-                    current_swap = int(_m.group()) if _m else 0
-                    _swap_disp = str(current_swap)
-                nf4_on = getattr(self, "quant_4bit_var", None) and self.quant_4bit_var.get()
-                if self._is_krea2_arch() and not nf4_on:
-                    messagebox.showwarning("Out of Memory",
-                        "CUDA ran out of memory during Krea 2 training.\n\n"
-                        "The biggest win on a smaller card is the 4-bit (NF4) Base toggle in the "
-                        "Memory & FP8 section: it shrinks the frozen base from ~14 GB to ~5.6 GB, so a "
-                        "full LoRA trains on a 10-12 GB card with no block swap.\n\n"
-                        "(Block swap helps too, but 4-bit frees far more. In-training previews on the "
-                        "Turbo need a bigger card — if they can't fit they auto-disable and training "
-                        "continues; evaluate the saved LoRA in ComfyUI.)")
-                else:
-                    messagebox.showwarning("Out of Memory",
-                        f"CUDA ran out of memory during training.\n\n"
-                        f"Current Block Swap: {_swap_disp}\n\n"
-                        f"Try increasing Block Swap on the Training tab "
-                        f"(Memory & FP8 section) to move more blocks to CPU. "
-                        f"If set to Auto, switch to a manual value like "
-                        f"{min(current_swap + 4, 16)}.")
-
-    def _browse_context_lora(self):
-        """File picker for the Context LoRA, filtered to .safetensors."""
-        path = filedialog.askopenfilename(
-            title="Select Context LoRA",
-            filetypes=[("SafeTensors", "*.safetensors"), ("All files", "*.*")],
-            initialdir=self._lora_initialdir(),
-        )
-        if path:
-            self.entries["CONTEXT_LORA_PATH"].delete(0, tk.END)
-            self.entries["CONTEXT_LORA_PATH"].insert(0, path)
-
-    def browse_file(self, setting_name, input_type):
-        # Resume Training points at a saved state dir, which lives under the LoRA
-        # output folder — open the Browse there so users don't hunt for it.
-        initial = self._pref_initialdir("lora_output_dir") if setting_name == "RESUME_TRAINING" else ""
-        if input_type == "directory":
-            path = filedialog.askdirectory(initialdir=initial)
-        else:
-            path = filedialog.askopenfilename(initialdir=initial)
-        if path:
-            self.settings[setting_name] = path
-            self.entries[setting_name].delete(0, tk.END)
-            self.entries[setting_name].insert(0, self.settings[setting_name])
-
-    def _tidy_lora_name(self):
-        """Clean the LoRA Name field in place. Returns (name, error or None).
-
-        The name becomes a filename, but not until the FIRST CHECKPOINT SAVE — an epoch in. A
-        stray character (a newline off a paste is the common one) trained for sixteen minutes
-        and then died inside safetensors with a bare OS error that named neither the setting nor
-        the character; and since the LoRA is written before the state dir, there was nothing left
-        to resume from (#70).
-
-        What has one obvious intent is fixed silently — surrounding whitespace, control
-        characters, trailing dots Windows discards anyway — and WRITTEN BACK to the widget, so
-        the field, the preset that gets persisted, the queue entry and --output_name cannot
-        disagree about what this run is called. Everything else is refused by name.
-        """
-        entry = self.entries.get("LORA_NAME")
-        raw = entry.get() if entry is not None else ""
-        name = "".join(c for c in raw if c >= " ").strip().rstrip(".").strip()
-        if entry is not None and name != raw:
-            entry.delete(0, tk.END)
-            entry.insert(0, name)
-        if not name:
-            return name, "LoRA name cannot be empty"
-        bad = next((c for c in name if c in '<>:"|?*/\\'), None)
-        if bad is not None:
-            return name, (f"LoRA name cannot contain '{bad}' — file names can't include that "
-                          f"character. Use letters, numbers, spaces, - _ or .")
-        return name, None
-
-    def validate_inputs(self):
-        """Validate all inputs before starting training"""
-        errors = []
-
-        # Get current architecture config
-        arch = self.architecture_var.get()
-        config = ARCHITECTURES.get(arch, ARCHITECTURES["Flux 2 Klein Base 9B"])
-
-        # Free-text numeric fields: a bare int()/float() further down the launch path used
-        # to raise inside a Tk callback — swallowed to stderr, invisible under the windowed
-        # launcher, and the Start button just "did nothing forever". Validate them HERE with
-        # a message naming the field. Batch Size 0/blank was the sharpest: it reached
-        # math.ceil(len(bucket)/batch_size) deep in the dataloader minutes after launch.
-        def _check_num(label, raw, cast, minimum=None):
-            raw = str(raw).strip()
-            try:
-                v = cast(raw)
-            except (TypeError, ValueError):
-                errors.append(f"{label} must be a number (got {raw!r})")
-                return
-            if minimum is not None and v < minimum:
-                errors.append(f"{label} must be at least {minimum} (got {raw})")
-
-        # Learning Rate box is ignored (and greyed) while Adaptive LR is on — don't let a
-        # stale value in a disabled box block Start.
-        if not (hasattr(self, 'adaptive_lr_var') and self.adaptive_lr_var.get()):
-            _check_num("Learning Rate", self.entries["LEARNING_RATE"].get(), float, 0)
-        _check_num("Network Dim (Rank)", self.entries["NETWORK_DIM"].get(), int, 1)
-        _check_num("Network Alpha", self.entries["NETWORK_ALPHA"].get(), float, 0)
-        if self._network_type_is_lokr():
-            _check_num("LoKR Factor", self.entries["LOKR_FACTOR"].get(), int, 2)
-        # Blocks to Train is free text, so a typo is caught HERE rather than after the 21 GB base
-        # has streamed in — and a queued run must never fail an hour later on a bad spec.
-        if self._is_minimax_arch():
-            _spec = minimax_block_spec(self.entries["MINIMAX_BLOCKS"].get())
-            # Likeness mode ignores (and disables) the box — a stale typo in it must not
-            # block the launch.
-            if self.entries["MINIMAX_LIKENESS_OPT"].get():
-                _spec = "all"
-            if _spec.lower() != "all":
-                try:
-                    from fizgig.minimax.trainer import parse_block_spec
-                    parse_block_spec(_spec, MINIMAX_NUM_BLOCKS)
-                except ValueError as e:
-                    errors.append(f"Blocks to Train: {e}")
-                except ImportError:
-                    pass
-            if minimax_lownoise_to_shift(self.entries["MINIMAX_LOWNOISE_PCT"].get()) is None:
-                errors.append("Low-noise training must be a number above 0 and below 100 "
-                              f"(got {self.entries['MINIMAX_LOWNOISE_PCT'].get()!r})")
-            _slow_spec = str(self.entries["MINIMAX_SLOW_BLOCKS"].get() or "").strip()
-            if _slow_spec:
-                try:
-                    from fizgig.minimax.trainer import parse_block_spec
-                    parse_block_spec(_slow_spec, MINIMAX_NUM_BLOCKS)
-                except ValueError as e:
-                    errors.append(f"Slower LR for blocks: {e}")
-                except ImportError:
-                    pass
-                _check_num("Slower LR multiplier",
-                           self.entries["MINIMAX_SLOW_LR_SCALE"].get(), float, 0)
-        _check_num("Max Train Epochs", self.entries["MAX_TRAIN_EPOCHS"].get(), int, 1)
-        _check_num("Save Every N Epochs", self.entries["SAVE_EVERY_N_EPOCHS"].get(), int, 1)
-        _check_num("Seed", self.entries["SEED"].get(), int)
-        _check_num("LoRA+ LR Ratio", self.entries["LORA_LR_RATIO"].get(), int, 1)
-        _check_num("Gradient Accumulation", self.entries["GRADIENT_ACCUMULATION"].get(), int, 1)
-        _check_num("Max Grad Norm", self.entries["MAX_GRAD_NORM"].get(), float, 0)
-        _check_num("Network Dropout", self.entries["NETWORK_DROPOUT"].get(), float, 0)
-        _check_num("Batch Size (Dataset)", self.dataset_batch_size_var.get(), int, 1)
-        # An unparseable megapixels value makes the TOML auto-saver skip its rewrite
-        # SILENTLY (#98 follow-up) — catch it here with a named error instead of the
-        # launch-time stale-config refusal.
-        _check_num("Target Megapixels (Dataset)", self.dataset_megapixels_var.get(), float, 0)
-        if "KEEP_LAST_N_STATES" in self.entries:
-            _check_num("Keep Last (states)", self.entries["KEEP_LAST_N_STATES"].get(), int, 1)
-
-        # Check required paths exist (sources: prefs_vars for model paths, hidden var for dataset)
-        dataset_config = self._get_path("DATASET_CONFIG")
-        if not dataset_config:
-            errors.append("Dataset config file path is empty — set the training image folder on the Start tab")
-        elif not os.path.exists(dataset_config):
-            errors.append(f"Dataset config file does not exist: {dataset_config}")
-
-        if config.get("is_minimax"):
-            # MiniMax H3 reads its three model paths from Preferences (minimax_*). All are
-            # required: the DiT to train (pruned int8 or bf16), the video VAE + Qwen3-VL-32B TE.
-            for pref_key, label in (
-                ("minimax_dit", "MiniMax H3 DiT"),
-                ("minimax_vae", "MiniMax H3 Video VAE"),
-                ("minimax_text_encoder", "Qwen3-VL-32B text encoder"),
-            ):
-                path = self._krea2_pref(pref_key)
-                if not path:
-                    errors.append(f"{label} path is empty (set it on the Preferences tab)")
-                elif not os.path.exists(path):
-                    errors.append(f"{label} file does not exist: {path}")
-            # Training Base = ref2va needs the ref2va model actually set — without this check
-            # the command builder would silently fall back to fl2va, the one thing the user
-            # explicitly asked it not to train on.
-            if (getattr(self, "minimax_train_base_var", None)
-                    and minimax_train_base(self.minimax_train_base_var.get()) == "ref2va"
-                    and not self._krea2_pref("minimax_ref_dit")):
-                errors.append("Training Base is set to Reference (ref2va) but 'DiT (reference)' "
-                              "is empty on the Preferences tab. Easiest fix: on the Preferences "
-                              "tab tick 'Include the reference DiT (+21 GB)' and hit "
-                              "'⬇ Download models for me' — it fetches the model and fills the "
-                              "path in for you. Or switch Training Base back to First/last "
-                              "frame.")
-            # Reference distillation needs the ref2va model and a real reference photo.
-            if getattr(self, "minimax_distill_var", None) and self.minimax_distill_var.get():
-                if not self._krea2_pref("minimax_ref_dit"):
-                    errors.append("Reference distillation needs the ref2va DiT — set "
-                                  "'DiT (reference)' on the Preferences tab. It is a different "
-                                  "model from the one above and the only H3 build that takes "
-                                  "reference images.")
-                _check_num("References each", self.entries["MINIMAX_DISTILL_REFS"].get(), int, 1)
-            # Multi Concept: each extra folder becomes its own [[datasets]] block, so it has to
-            # exist, be distinct, and carry its own captions. The dataset layer refuses two
-            # blocks sharing a cache_directory, and the cache path hashes a case-folded,
-            # slash-stripped path — so C:\A and c:/a/ are the SAME folder as far as it cares.
-            if (getattr(self, "minimax_multiconcept_var", None)
-                    and self.minimax_multiconcept_var.get()):
-                _seen = {self.image_folder_var.get().strip().lower()
-                         .replace("\\", "/").rstrip("/")}
-                _extra = [v.get().strip() for v in getattr(self, "_concept_folder_vars", [])]
-                if not any(_extra):
-                    errors.append("Multi Concept is on but no second subject folder is set — "
-                                  "pick one, or turn the mode off.")
-                for _f in _extra:
-                    if not _f:
-                        continue
-                    _norm = _f.lower().replace("\\", "/").rstrip("/")
-                    if _norm in _seen:
-                        errors.append(f"Multi Concept: {_f} is the same folder as another "
-                                      f"subject — each needs its own folder.")
-                        continue
-                    _seen.add(_norm)
-                    if not os.path.isdir(_f):
-                        errors.append(f"Multi Concept: folder does not exist: {_f}")
-                        continue
-                    _ext = (self.dataset_caption_ext_var.get().strip() or ".txt")
-                    if not any(fn.lower().endswith(_ext.lower()) for fn in os.listdir(_f)):
-                        errors.append(
-                            f"Multi Concept: no {_ext} captions in {_f}. Caption both folders "
-                            f"before training — each subject needs its own trigger word in "
-                            f"every caption, or they will blend.")
-        elif config.get("is_krea2"):
-            # Krea 2 reads its own four model paths from Preferences (krea2_*). The
-            # Turbo DiT is only required when in-training previews are enabled.
-            krea2_required = [
-                ("krea2_raw_dit", "Krea 2 RAW DiT"),
-                ("krea2_vae", "Qwen-Image VAE"),
-                ("krea2_text_encoder", "Qwen3-VL-4B text encoder"),
-            ]
-            if self.sample_enabled_var.get():
-                # raw_lora engine renders previews on the training DiT + Turbo LoRA — the LoRA
-                # is auto-downloaded at training start if missing, so nothing is hard-required
-                # here (a failed download degrades to no previews, with console messages, never
-                # a blocked run). The classic engine still needs the Turbo checkpoint.
-                if self._krea2_preview_engine() != "raw_lora":
-                    krea2_required.append(("krea2_turbo_dit", "Krea 2 Turbo DiT (fp8) — needed for previews"))
-            for pref_key, label in krea2_required:
-                path = self._krea2_pref(pref_key)
-                if not path:
-                    errors.append(f"{label} path is empty (set it on the Preferences tab)")
-                elif not os.path.exists(path):
-                    errors.append(f"{label} file does not exist: {path}")
-        else:
-            vae_model = self._get_path("VAE_MODEL")
-            if not vae_model:
-                errors.append("VAE model file path is empty (set on the Preferences tab)")
-            elif not os.path.exists(vae_model):
-                errors.append(f"VAE model file does not exist: {vae_model}")
-
-            dit_model = self._get_path("DIT_MODEL")
-            if not dit_model:
-                errors.append("DiT model file path is empty (set on the Preferences tab)")
-            elif not os.path.exists(dit_model):
-                errors.append(f"DiT model file does not exist: {dit_model}")
-
-            # Architecture-specific validation (T5/CLIP are dead for Klein but kept for future flexibility)
-            if config["uses_t5"]:
-                t5_model = self._get_path("T5_MODEL")
-                if not t5_model:
-                    errors.append("T5 model file path is empty")
-                elif not os.path.exists(t5_model):
-                    errors.append(f"T5 model file does not exist: {t5_model}")
-
-            if config["uses_text_encoder"]:
-                text_encoder = self._get_path("TEXT_ENCODER")
-                if not text_encoder:
-                    errors.append("Text encoder file path is empty (set on the Preferences tab)")
-                elif not os.path.exists(text_encoder):
-                    errors.append(f"Text encoder file does not exist: {text_encoder}")
-
-            if config["uses_clip"]:
-                clip_model = self._get_path("CLIP_MODEL")
-                if not clip_model:
-                    errors.append("CLIP model file path is empty")
-                elif not os.path.exists(clip_model):
-                    errors.append(f"CLIP model file does not exist: {clip_model}")
-
-        # Validate numeric fields. With Adaptive LR on, the Learning Rate box is IGNORED
-        # (the run starts at the geometric midpoint of Min/Max), so only Min < Max matters
-        # — the old "starting LR exceeds Max" check no longer applies.
-        _adaptive_on = hasattr(self, 'adaptive_lr_var') and self.adaptive_lr_var.get()
-        if _adaptive_on:
-            try:
-                max_lr_str = self.entries["ADAPTIVE_LR_MAX"].get().split(" ")[0]
-                min_lr_str = self.entries["ADAPTIVE_LR_MIN"].get().split(" ")[0]
-                if float(min_lr_str) >= float(max_lr_str):
-                    errors.append(f"Adaptive Min LR ({min_lr_str}) must be lower than Max LR ({max_lr_str}).")
-            except (ValueError, KeyError):
-                errors.append("Adaptive Min/Max LR must be valid numbers.")
-        else:
-            try:
-                lr = float(self.entries["LEARNING_RATE"].get())
-                if lr <= 0:
-                    errors.append("Learning rate must be positive")
-            except ValueError:
-                errors.append("Learning rate must be a valid number")
-
-        # Context LoRA validation (supported by both Klein and Krea 2).
-        ctx_path = self.entries.get("CONTEXT_LORA_PATH").get().strip() if "CONTEXT_LORA_PATH" in self.entries else ""
-        if ctx_path:
-            if not os.path.exists(ctx_path):
-                errors.append(f"Context LoRA file does not exist: {ctx_path}")
-            elif not ctx_path.lower().endswith(".safetensors"):
-                errors.append(f"Context LoRA must be a .safetensors file: {ctx_path}")
-            try:
-                ctx_strength = float(self.entries["CONTEXT_LORA_STRENGTH"].get())
-                if not (0.0 <= ctx_strength <= 2.0):
-                    errors.append(f"Context LoRA Strength ({ctx_strength}) must be between 0.0 and 2.0")
-            except (ValueError, KeyError):
-                errors.append("Context LoRA Strength must be a valid number")
-
-        try:
-            network_dim = int(self.entries["NETWORK_DIM"].get())
-            if network_dim <= 0:
-                errors.append("Network dim must be a positive integer")
-        except ValueError:
-            errors.append("Network dim must be a valid integer")
-
-        try:
-            network_alpha = float(self.entries["NETWORK_ALPHA"].get())
-            if network_alpha < 0:
-                errors.append("Network alpha must be non-negative")
-        except ValueError:
-            errors.append("Network alpha must be a valid number")
-
-        try:
-            epochs = int(self.entries["MAX_TRAIN_EPOCHS"].get())
-            if epochs <= 0:
-                errors.append("Max train epochs must be a positive integer")
-        except ValueError:
-            errors.append("Max train epochs must be a valid integer")
-
-        try:
-            save_epochs = int(self.entries["SAVE_EVERY_N_EPOCHS"].get())
-            if save_epochs <= 0:
-                errors.append("Save every N epochs must be a positive integer")
-            # Per-category retirement epoch: blank = never, else a positive whole number.
-            _rk = self.entries.get("MIXED_STOP_EPOCH")
-            _rv = str(_rk.get() if _rk else "").strip()
-            if _rv and (not _rv.isdigit() or int(_rv) <= 0):
-                errors.append(f"'Finish one category early: after epoch' must be blank or "
-                              f"a positive whole number, not {_rv!r}")
-        except ValueError:
-            errors.append("Save every N epochs must be a valid integer")
-
-        try:
-            blocks_swap = self._parse_blocks_swap()
-            if blocks_swap < 0:
-                errors.append("Blocks swap must be non-negative")
-            elif blocks_swap > config["blocks_swap_max"]:
-                errors.append(f"Blocks swap ({blocks_swap}) exceeds maximum for {arch} ({config['blocks_swap_max']})")
-        except ValueError:
-            errors.append("Blocks swap must be a valid integer")
-
-        _name, _name_error = self._tidy_lora_name()
-        if _name_error:
-            errors.append(_name_error)
-
-        # Check output directory
-        output_dir = self.entries["LORA_OUTPUT_DIR"].get()
-        if not output_dir:
-            errors.append("LoRA output directory is empty")
-
-        # Check resume path if specified
-        resume_path = self.entries["RESUME_TRAINING"].get()
-        if resume_path and resume_path.strip() and not os.path.exists(resume_path):
-            errors.append(f"Resume training path does not exist: {resume_path}")
-
-        # Check caption files exist in the dataset folder
-        image_dir = self.image_folder_var.get().strip()
-        caption_ext = self.dataset_caption_ext_var.get().strip()
-        # A SET folder that no longer exists is an error, not a skip: the TOML regenerator
-        # early-returns on a missing folder, so proceeding trains whatever dataset the TOML
-        # last pointed at — silently, under this run's name.
-        if image_dir and not os.path.isdir(image_dir):
-            errors.append(f"Training image folder does not exist: {image_dir}")
-        if image_dir and os.path.isdir(image_dir) and caption_ext:
-            import glob as _glob
-            # glob.escape is load-bearing here: a folder like "[subject] photos" made this
-            # find zero captions and block training with "No caption files found", while the
-            # Captions tab (os.listdir) read and wrote that same folder perfectly happily.
-            caption_files = _glob.glob(os.path.join(_glob.escape(image_dir), "*" + caption_ext))
-            if not caption_files:
-                errors.append(
-                    f"No caption files (*{caption_ext}) found in {image_dir}. "
-                    f"Use the Captions tab to generate them first."
-                )
-
-        if errors:
-            error_message = "Please fix the following issues:\n\n" + "\n".join(f"• {e}" for e in errors)
-            messagebox.showerror("Validation Error", error_message)
-            return False
-
-        return True
-
-    @staticmethod
-    def _pipeline_exit_routes_to_state_machine(name, returncode):
-        """Which subprocess exits the pause/queue state machine must hear about.
-
-        Training exits: always (clean finish, failure, stop, pause all carry state).
-        Caching phases ("Cache Preparation" / "Text Encoder Caching"): only NONZERO exits —
-        a clean cache exit continues the pipeline via its callback and the run is still
-        alive. Before this predicate existed, a failed or stopped caching phase left
-        training_state stranded at "running" with no process: the Start button read
-        "Queue Train" but launched, Pause pointed at nothing, the queue window pinned a
-        dead run, and a queued job that died in caching vanished without the HELD notice.
-        """
-        n = (name or "").lower()
-        if "training" in n:
-            return True
-        return "cach" in n and returncode != 0
 
     def run_subprocess(self, cmd, name, callback=None):
         """Run a subprocess and handle its output with UTF-8 encoding"""
